train

+++ test/read_mnist_data2.py
@@ -29,7 +29,7 @@ def maybe_download(filename, work_directory):
 
 
 def _read32(bytestream):
-    dt = numpy.dtype(numpy.float32).newbyteorder('>')
+    dt = numpy.dtype(numpy.uint32).newbyteorder('>')
     return numpy.frombuffer(bytestream.read(4), dtype=dt)[0]
 
 
@@ -46,7 +46,7 @@ def extract_images(filename):
         rows = _read32(bytestream)
         cols = _read32(bytestream)
         buf = bytestream.read(rows * cols * num_images)
-        data = numpy.frombuffer(buf, dtype=numpy.uint8)
+        data = numpy.frombuffer(buf, dtype=numpy.float32)
         data = data.reshape(num_images, rows, cols, 1)
         return data
